bosminer: Account generated wok as an 64bit value
- add trait 'stats::AtomicCounter' to support generic 'stats::Counter'
- convert 'stats::Counter' to generic implementation
- add macro 'stats::atomic_counter_impl' to generate implementation of
  trait 'stats::AtomicCounter' for standard 'std::sync::atomic' struts
- add aliases to useful types 'CounterU64', 'CounterUsize'
- change return value of methods 'generated_work' to 'CounterU64'

+++ open/bosminer/bosminer/src/work/solver.rs
@@ -155,7 +155,7 @@ impl Generator {
                 }
             };
             // determine how much work has been generated for current work assignment
-            let work_amount = work.generated_work_amount();
+            let work_amount = work.generated_work_amount() as u64;
             work.origin()
                 .client_stats()
                 .generated_work()
